ci(build): -Werror first-party 库构建 + 13 处 cosmetic 告警清零（S12-7 闭环）
S12-7 构建加固最后一项：把 first-party 告警升为错误作新告警回归护栏。
13 处实际告警（原估「~15」含 vendored 头误算）均为零行为风险的机械修复，
third_party 隔离后 -Werror 只作用于 first-party。

告警清零（6 文件，零行为变更）
* `-Wshadow` ×7：inverted.hpp 的 `max_tf` → `blk_max_tf`（两处构造块）；
  bm25/inverted.cpp bool_search 的 `k` → `mk`（与外层 `k=` 参数避免遮蔽）；
  deserialize 的 `pos` → `tf_pos` / `dl_pos`（两段 VByte 解码独立命名）。
* `-Wsign-conversion` ×4：thread_pool.hpp 的 tbb 队列 `set_capacity` /
  `size()` 两次 `static_cast<std::ptrdiff_t>` / `std::size_t`；search_layer.cpp
  的 std::min(ranked.size(), k) 与 hnsw.cpp 两处 std::partial_sort
  `begin() + n` 的 ptrdiff_t 转换。
* `-Wunused-function` ×1：删除 cask.cpp 末尾未被调用的 `str_to_bytes` helper
  （只用 `bytes_to_view`，方向相反的局部便利函数早已被内联替代）。
* `-Wunused-parameter` ×1：keydir.cpp 的 `put_overwrite` 的 `key` 参数加
  `[[maybe_unused]]`（分支 (B)/(C) 不需要读 key，仅签名统一）。

机制
* 新增 `option(BITCASK_WERROR OFF)`（`CMakeLists.txt`）：开时给
  `bitcask_warnings` INTERFACE 加 `-Werror`。**默认关**——避免新编译器新告警
  破坏下游/本地构建。
* third_party 隔离：`cppjieba` INTERFACE 改 `SYSTEM` include；编译器视为
  系统头、不对其大量 `-Wold-style-cast` / `-Wsign-conversion` 告警报错。
  limonp 头经 cppjieba 传递亦同。
* 新 CI job `werror-lib`（`.github/workflows/ci.yml`）：ubuntu-24.04 + GCC 13
  （从 PPA 装）+ ccache + Release + `BITCASK_WERROR=ON` + `BUILD_TESTING=OFF` +
  只 `cmake --build --target bitcask_static bitcask_shared`——纯库构建、无测试/
  无 bench，护栏纯净。

验证
* 本地 `-DBITCASK_WERROR=ON` 建库（first-party only）0 错误 0 告警。
* 常规构建全量 486/486 零回归（13 处编辑行为中性）。
* ci.yml YAML 合法，5 jobs（gcc / clang / werror-lib / asan-ubsan / tsan）。

未做（有意）
* 常规 build（含 third_party）仍不开 -Werror——third_party 头控不住且与本批
  护栏目标无关。
* `-Werror=...` 分类（如 -Wno-error=shadow）不引入——按告警类去禁会绕过护栏
  价值。

TASK.md S12-7 子项（版本单一真源 + -Werror）全部 ✅。

+++ include/bitcask/thread_pool.hpp
@@ -155,7 +155,7 @@ class IndexTaskQueue {
 public:
     explicit IndexTaskQueue(std::size_t capacity = 10240)
     {
-        queue_.set_capacity(capacity);
+        queue_.set_capacity(static_cast<std::ptrdiff_t>(capacity));
     }
 
     // C1:tbb 队列内部以 fence 同步,TSan 不建模独立 fence——producer
@@ -180,7 +180,7 @@ public:
         return true;
     }
 
-    std::size_t size() const { return queue_.size(); }
+    std::size_t size() const { return static_cast<std::size_t>(queue_.size()); }
 
 private:
 #if defined(__SANITIZE_THREAD__) || \
